Add icons and use palette
Adds icons for the buttons and specifies colourings based on the
palette. The aim is to give a more consistent appearance across
platforms.

The change also updates the dom walking code so that a red shadow is
added to widgets during the walk rather than a border. This is to
prevent the items changing size, while still otherwise producing a
similar red border effect.