Set CMake policy QTP0004
Sets the CMake policy QTP0004 to NEW. That is for each directory with
QML files in a module a separate qmldir file is generated, rather than a
file just for the root directory of a module.

See:
https://doc.qt.io/qt-6/qt-cmake-policy-qtp0004.html

Setting the policy prevents a warning from being generated during
execution of CMake.